-> Edit job implemented.

+++ app/src/main/java/com/example/managerbookfreelancer/fragments/RecyclerViewJobsFragment.kt
@@ -2,12 +2,15 @@ package com.example.managerbookfreelancer.fragments
 
 import android.os.Bundle
 import android.view.*
+import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.MenuProvider
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.Lifecycle
+import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import com.example.managerbookfreelancer.R
 import com.example.managerbookfreelancer.adapter.AdapterListJobs
 import com.example.managerbookfreelancer.core.dataBase.JobAppDataBase
@@ -25,6 +28,7 @@ class RecyclerViewJobsFragment : Fragment() {
     private var _binding: FragmentRecyclerViewJobsBinding? = null
     private val binding get() = _binding!!
     private lateinit var adapter: AdapterListJobs
+    private lateinit var recyclerView: RecyclerView
     var showOldItens: Boolean = false
 
     private val viewModel: JobsViewModel by activityViewModels(
@@ -38,10 +42,30 @@ class RecyclerViewJobsFragment : Fragment() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        setDialogDelete()
+    }
+
+    private fun setDialogDelete() {
         adapter = AdapterListJobs(onClick = {
-            CoroutineScope(Dispatchers.IO).launch {
-                viewModel.delete(jobEntity = it)
-            }
+            val dialog = AlertDialog.Builder(requireContext())
+                .setCancelable(true)
+                .setTitle("Are your want delete or cancel this job:")
+                .setMessage("Professional: ${it.ownerName} Date: ${Resoucers.fromLongToString(it.weddingDay)}")
+                .setPositiveButton("Delete") { _, _ ->
+                    CoroutineScope(Dispatchers.IO).launch {
+                        viewModel.delete(jobEntity = it)
+                    }
+                }.setNeutralButton("Cancel"){ _, _ ->
+                }
+                .setNegativeButton("Edite") { _, _ ->
+                    val action = RecyclerViewJobsFragmentDirections.actionRecyclerViewJobsFragmentToFormNewJobFragment(
+                        it.idJob
+                    )
+                    findNavController().navigate(action)
+                }.create()
+
+            dialog.show()
+
         })
     }
 
@@ -51,25 +75,28 @@ class RecyclerViewJobsFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View {
 
-        // mostra a action bar
+        // Show action bar
         val activity = activity as AppCompatActivity?
         if (activity != null) {
             activity.supportActionBar?.show()
         }
 
         _binding = FragmentRecyclerViewJobsBinding.inflate(inflater, container, false)
+        recyclerView = binding.RCFragmentListJobs
+        recyclerView.adapter = adapter
+        recyclerView.layoutManager = LinearLayoutManager(requireContext())
+
+        // Register the Context menu on RecyclerView
+        registerForContextMenu(recyclerView)
+
+
         return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         setUpToolBar()
-
-        binding.RCFragmentListJobs.layoutManager = LinearLayoutManager(requireContext())
-        binding.RCFragmentListJobs.adapter = adapter
         observeJobs()
-
-
     }
 
     private fun observeJobs() {
